ml_project/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.0
pillow>=8.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.62.0
scikit-learn>=0.24.0
scipy>=1.5.0